# public/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API UKM UAKI - Backend Server",
    "description": "Dokumentasi API untuk sistem backend UKM UAKI. Pastikan untuk mengisi Bearer Token (kecuali untuk endpoint publik) sebelum melakukan request.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Masukkan token JWT yang didapat dari proses Login (Admin/Registrant)"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/admins/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login Admin",
        "description": "Endpoint login untuk Admin dan Super Admin. Akan menghasilkan token JWT dengan role 'admin'.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "admin@uaki.com" },
                  "password": { "type": "string", "example": "rahasia123" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Berhasil mendapatkan Token JWT Admin" } }
      }
    },
    "/registrants/auth/google": {
      "post": {
        "tags": ["Authentication"],
        "summary": "SSO Google Login & Auto Register Pendaftar",
        "description": "Endpoint login utama untuk pendaftar menggunakan akun Google (email @student.ub.ac.id). Jika email belum terdaftar di database, sistem akan otomatis membuatkan akun baru dengan status PENDING.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "google_token": { "type": "string", "description": "ID Token otentikasi yang digenerate oleh SDK Google OAuth di Frontend" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Berhasil login / membuat akun otomatis" } }
      }
    },
    "/registrants/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login Manual Pendaftar (Opsi Alternatif)",
        "description": "Hanya bisa digunakan jika akun tidak dibuat melalui SSO Google.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "pendaftar@student.ub.ac.id" },
                  "password": { "type": "string", "example": "pass123" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Berhasil mendapatkan Token JWT Registrant" } }
      }
    },
    "/admins": {
      "get": {
        "tags": ["Admins (Super Admin Only)"],
        "summary": "List Semua Admin",
        "description": "Menampilkan daftar seluruh admin. Mendukung pagination.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "Daftar admin berhasil diambil" } }
      },
      "post": {
        "tags": ["Admins (Super Admin Only)"],
        "summary": "Buat Akun Admin Baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" },
                  "is_super_admin": { "type": "boolean", "description": "Set true jika ingin memberikan akses super admin" },
                  "is_active": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Akun admin berhasil dibuat" } }
      }
    },
    "/admins/{admin_id}": {
      "get": {
        "tags": ["Admins (Super Admin Only)"],
        "summary": "Get Detail Admin",
        "parameters": [
          { "name": "admin_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Detail Admin" } }
      },
      "put": {
        "tags": ["Admins (Super Admin Only)"],
        "summary": "Update Admin",
        "parameters": [
          { "name": "admin_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string", "description": "Masukkan password baru jika ingin mengubah, jika tidak biarkan password lama" },
                  "is_super_admin": { "type": "boolean" },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Update admin sukses" } }
      },
      "delete": {
        "tags": ["Admins (Super Admin Only)"],
        "summary": "Hapus Admin",
        "parameters": [
          { "name": "admin_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Admin terhapus" } }
      }
    },
    "/registrants/me": {
      "get": {
        "tags": ["Registrants (Self)"],
        "summary": "Lihat Profil Sendiri",
        "description": "Mengambil data profil pendaftar berdasarkan JWT Token. Gunakan ini untuk mengecek status kelulusan dan melihat jadwal screening.",
        "responses": { "200": { "description": "Berhasil menampilkan profil" } }
      },
      "put": {
        "tags": ["Registrants (Self)"],
        "summary": "Update Form Pendaftaran Sendiri",
        "description": "Endpoint bagi pendaftar untuk melengkapi form pendaftaran. Semua isian bersifat string bebas, kecuali untuk division_1 dan division_2.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "nickname": { "type": "string", "description": "Nama Panggilan" },
                  "nim": { "type": "string", "description": "Nomor Induk Mahasiswa (Opsional saat auto-register)" },
                  "angkatan": { "type": "string" },
                  "prodi": { "type": "string" },
                  "fakultas": { "type": "string" },
                  "domicile": { "type": "string", "description": "Alamat Domisili / Tempat Tinggal di Malang" },
                  "origin_city": { "type": "string", "description": "Asal Daerah / Kota Asal" },
                  "school_origin": { "type": "string", "description": "Asal Sekolah (SMA/SMK)" },
                  "has_rohis_exp": { "type": "boolean", "description": "Pernah mengikuti Rohis di SMA (true/false)" },
                  "phone": { "type": "string", "description": "Nomor WhatsApp aktif" },
                  "division_1": { "type": "string", "enum": ["KP", "MENTORING", "SYIAR", "CM", "HUMAS", "MUCC", "RAB", "EKRAF", "PPM"], "description": "Pilihan divisi prioritas pertama" },
                  "division_2": { "type": "string", "enum": ["KP", "MENTORING", "SYIAR", "CM", "HUMAS", "MUCC", "RAB", "EKRAF", "PPM"], "description": "Pilihan divisi cadangan" },
                  "swot_s": { "type": "string", "description": "Kekuatan (Strengths)" },
                  "swot_w": { "type": "string", "description": "Kelemahan (Weaknesses)" },
                  "swot_o": { "type": "string", "description": "Peluang (Opportunities)" },
                  "swot_t": { "type": "string", "description": "Ancaman (Threats)" },
                  "organization_exp": { "type": "string", "description": "Riwayat Organisasi" },
                  "commitment": { "type": "string", "description": "Deskripsi komitmen di UAKI" },
                  "cv_url": { "type": "string", "description": "URL dari endpoint upload CV" },
                  "twibbon_url": { "type": "string", "description": "URL dari endpoint upload Twibbon" },
                  "portofolio_url": { "type": "string", "description": "URL dari endpoint upload Portofolio (Khusus Departemen CM)" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Formulir berhasil diperbarui" } }
      }
    },
    "/registrants/file": {
      "post": {
        "tags": ["Registrants (Self)"],
        "summary": "Upload Berkas (CV, Twibbon, Portofolio)",
        "description": "Digunakan oleh pendaftar untuk mengunggah CV (.pdf), Twibbon (.jpg/.png/.webp), dan/atau Portofolio (.pdf). Dapat mengirim salah satu atau ketiga file sekaligus dalam satu request.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "cv": { "type": "string", "format": "binary", "description": "Format file .pdf, maksimal 10 MB" },
                  "twibbon": { "type": "string", "format": "binary", "description": "Format file .jpg, .jpeg, .png, .webp, maksimal 5 MB" },
                  "portofolio": { "type": "string", "format": "binary", "description": "Format file .pdf, maksimal 10 MB (Khusus Departemen CM)" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Berkas sukses diupload, mengembalikan cv_url, twibbon_url, dan/atau portofolio_url" } }
      }
    },
    "/registrants/export": {
      "get": {
        "tags": ["Registrants (Admin)"],
        "summary": "Export Data Pendaftar (.xlsx)",
        "description": "Mengunduh seluruh data pendaftar dalam format file spreadsheet Excel (.xlsx). Mendukung filter opsional berdasarkan status dan divisi.",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Filter pencarian nama/email/nim/prodi/fakultas" },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ALL", "PENDING", "LOLOS_BERKAS", "DITERIMA", "DITOLAK"], "default": "ALL" } },
          { "name": "division", "in": "query", "schema": { "type": "string" }, "description": "Filter divisi (contoh: CM, EKRAF, HUMAS, dll)" }
        ],
        "responses": {
          "200": {
            "description": "File Excel (.xlsx) berhasil di-generate dan diunduh",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/registrants": {
      "get": {
        "tags": ["Registrants (Admin)"],
        "summary": "List Pendaftar",
        "description": "Menampilkan daftar seluruh pendaftar yang sudah mengirimkan formulir pendaftaran. Akses dibatasi untuk role admin. Mendukung pagination dan filter status.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ALL", "PENDING", "LOLOS_BERKAS", "DITERIMA", "DITOLAK"], "default": "ALL" } }
        ],
        "responses": { "200": { "description": "List data pendaftar berhasil diambil" } }
      }
    },
    "/registrants/{registrant_id}": {
      "get": {
        "tags": ["Registrants (Admin)"],
        "summary": "Detail Data Pendaftar",
        "parameters": [
          { "name": "registrant_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Detail data ditampilkan" } }
      },
      "delete": {
        "tags": ["Registrants (Admin)"],
        "summary": "Hapus Pendaftar",
        "description": "Menghapus data pendaftar beserta file CV fisiknya yang ada di storage.",
        "parameters": [
          { "name": "registrant_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Pendaftar berhasil dihapus" } }
      }
    },
    "/registrants/{registrant_id}/status": {
      "patch": {
        "tags": ["Registrants (Admin)"],
        "summary": "Update Status Kelulusan",
        "description": "Ubah status kelulusan pendaftar.",
        "parameters": [
          { "name": "registrant_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": { 
                    "type": "string", 
                    "enum": ["PENDING", "LOLOS_BERKAS", "DITERIMA", "DITOLAK"],
                    "description": "Status kelulusan terkini"
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Status berhasil diubah" } }
      }
    },
    "/registrants/send-schedule": {
      "post": {
        "tags": ["Registrants (Admin)"],
        "summary": "Kirim Jadwal via Email",
        "description": "Mengirimkan email jadwal screening secara massal ke pendaftar yang dipilih. API berjalan secara background (Asynchronous) sehingga me-return respons 200 seketika.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "registrant_ids": {
                    "type": "array",
                    "description": "List ID pendaftar yang ingin dikirimi jadwal",
                    "items": { "type": "string", "format": "uuid" }
                  },
                  "screening_date": { 
                    "type": "string", 
                    "description": "Waktu screening. Format WAJIB: YYYY-MM-DD HH:MM:SS",
                    "example": "2026-08-30 09:00:00" 
                  },
                  "screening_location": { 
                    "type": "string", 
                    "description": "Lokasi fisik (misal: Ruang 1 Filkom)",
                    "example": "Gedung A Filkom" 
                  },
                  "screening_link": { 
                    "type": "string", 
                    "description": "Link URL meeting jika online",
                    "example": "https://meet.google.com/xyz-abc-def" 
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Proses background pengiriman email berjalan" } }
      }
    },
    "/articles": {
      "get": {
        "tags": ["Articles"],
        "summary": "List Artikel (Publik)",
        "security": [],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "Daftar artikel berhasil diambil" } }
      },
      "post": {
        "tags": ["Articles (Admin)"],
        "summary": "Buat Artikel Baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": { "type": "string", "enum": ["islam", "umum", "informasi"], "description": "Kategori sesuai tipe ENUM di database" },
                  "title": { "type": "string" },
                  "slug": { "type": "string", "description": "Slug URL SEO-friendly (Opsional, otomatis di-generate jika kosong)" },
                  "value": { "type": "string", "description": "Isi/Body artikel dalam bentuk string HTML atau Markdown" },
                  "img_url": { "type": "string", "description": "URL gambar header artikel" },
                  "is_active": { "type": "boolean", "default": true }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Artikel dibuat" } }
      }
    },
    "/articles/image": {
      "post": {
        "tags": ["Articles (Admin)"],
        "summary": "Upload Gambar Artikel",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary", "description": "File gambar untuk artikel (png/jpg)" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Gambar terupload, me-return URL gambar" } }
      }
    },
    "/articles/{article_id}": {
      "get": {
        "tags": ["Articles"],
        "summary": "Detail Artikel (Publik)",
        "description": "Mengambil detail artikel berdasarkan UUID atau Slug URL (contoh: /articles/tips-sukses-kuliah).",
        "security": [],
        "parameters": [
          { "name": "article_id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "UUID atau Slug Artikel" }
        ],
        "responses": { "200": { "description": "Detail artikel" } }
      },
      "put": {
        "tags": ["Articles (Admin)"],
        "summary": "Update Artikel",
        "parameters": [
          { "name": "article_id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "UUID atau Slug Artikel" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "category": { "type": "string", "enum": ["islam", "umum", "informasi"] },
                  "title": { "type": "string" },
                  "slug": { "type": "string", "description": "Slug URL kustom baru (Opsional)" },
                  "value": { "type": "string" },
                  "img_url": { "type": "string" },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Artikel berhasil diperbarui" } }
      },
      "delete": {
        "tags": ["Articles (Admin)"],
        "summary": "Hapus Artikel",
        "parameters": [
          { "name": "article_id", "in": "path", "required": true, "schema": { "type": "string" }, "description": "UUID atau Slug Artikel" }
        ],
        "responses": { "200": { "description": "Artikel terhapus" } }
      }
    },
    "/media-categories": {
      "get": {
        "tags": ["Media Categories"],
        "summary": "List Kategori Media (Publik)",
        "security": [],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "Daftar kategori berhasil diambil" } }
      },
      "post": {
        "tags": ["Media Categories (Super Admin Only)"],
        "summary": "Buat Kategori Media Baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "description": "Nama kategori media (misal: Dokumentasi Kegiatan)" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Kategori Dibuat" } }
      }
    },
    "/media-categories/{media_category_id}": {
      "get": {
        "tags": ["Media Categories"],
        "summary": "Detail Kategori Media (Publik)",
        "security": [],
        "parameters": [
          { "name": "media_category_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Detail Kategori" } }
      },
      "put": {
        "tags": ["Media Categories (Super Admin Only)"],
        "summary": "Update Kategori",
        "parameters": [
          { "name": "media_category_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Update kategori berhasil" } }
      },
      "delete": {
        "tags": ["Media Categories (Super Admin Only)"],
        "summary": "Hapus Kategori",
        "parameters": [
          { "name": "media_category_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Hapus berhasil" } }
      }
    },
    "/media": {
      "get": {
        "tags": ["Media"],
        "summary": "List Media (Publik)",
        "security": [],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "Daftar media berhasil diambil" } }
      },
      "post": {
        "tags": ["Media (Admin)"],
        "summary": "Buat Media Baru",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "img_url": { "type": "string" },
                  "category_id": { "type": "integer", "description": "ID kategori media dari endpoint Kategori" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Media berhasil dibuat" } }
      }
    },
    "/media/image": {
      "post": {
        "tags": ["Media (Admin)"],
        "summary": "Upload Gambar Media",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary", "description": "File gambar" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Upload sukses, me-return URL" } }
      }
    },
    "/media/{media_id}": {
      "get": {
        "tags": ["Media"],
        "summary": "Detail Media (Publik)",
        "security": [],
        "parameters": [
          { "name": "media_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Detail media" } }
      },
      "put": {
        "tags": ["Media (Admin)"],
        "summary": "Update Media",
        "parameters": [
          { "name": "media_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "img_url": { "type": "string" },
                  "category_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Update media berhasil" } }
      },
      "delete": {
        "tags": ["Media (Admin)"],
        "summary": "Hapus Media",
        "parameters": [
          { "name": "media_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": { "200": { "description": "Media terhapus" } }
      }
    }
  }
}